// generated/src/app/api-models/smproduct-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmproductS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  productOwnerid: number;
  productOwnertype: number;
  productProductid: number;
  productDependencies: string;
  productSharable: string;
  productSelected: string;
  productPrdfile: string;
}